Add Season class and tests
